implementation for local miner calls

+++ miner/state/block_chain_state_test.go
@@ -239,7 +239,7 @@ func TestAccountConfig(t *testing.T) {
 			Filename: filenames[0],
 			Data: datum[0],
 			Creator: strconv.Itoa(1),
-			RecordNumber: uint32(0),
+			RecordNumber: uint16(0),
 		}
 		records[0] = &record
 		ee := crypto.BlockElement{
@@ -294,7 +294,7 @@ func TestAccountConfig(t *testing.T) {
 		// Add append node
 		AddAppendNode(tree)
 
-		// Strictly speaking the appendFee should always == 1, but for testing purposes we set it to something
+		// Strictly speaking the AppendFee should always == 1, but for testing purposes we set it to something
 		// larger here
 		bkState, err := NewAccountsState(10, createFee, opReward, noOpReward, tree.GetLongestChain())
 		if err != nil {
