+++ miner/state/file_system_state.go
@@ -11,7 +11,7 @@ type Filename string
 type FileData []byte
 type FileInfo struct {
 	Creator string
-	NumberOfRecords uint32
+	NumberOfRecords uint16
 	Data    FileData
 }
